Refactor - create package, move ruleset evaluation utils to _utils
