+++ awswrangler/data_quality/__init__.py
@@ -1,0 +1,7 @@
+"""Glue Data Quality Module."""
+
+from awswrangler.data_quality._create import create_ruleset
+
+__all__ = [
+    "create_ruleset",
+]
